report average TPS in CryptoBench tests; remove CryptoBenchMerkleDb
