+++ platform-sdk/swirlds-benchmarks/src/jmh/java/com/swirlds/benchmark/CryptoBench.java
@@ -35,7 +35,7 @@ import org.openjdk.jmh.annotations.Warmup;
 @State(Scope.Thread)
 @Warmup(iterations = 0)
 @Measurement(iterations = 1)
-public abstract class CryptoBench extends VirtualMapBench {
+public class CryptoBench extends VirtualMapBench {
 
     private static final Logger logger = LogManager.getLogger(CryptoBench.class);
 
@@ -51,6 +51,12 @@ public abstract class CryptoBench extends VirtualMapBench {
     private static final int FIXED_KEY_ID2 = 1;
     private Bytes fixedKey1;
     private Bytes fixedKey2;
+
+    @Override
+    public void beforeTest(String name) {
+        super.beforeTest(name);
+        updateMerkleDbPath();
+    }
 
     @Override
     String benchmarkName() {
@@ -101,6 +107,15 @@ public abstract class CryptoBench extends VirtualMapBench {
         tps.set(average(delta));
     }
 
+    private void totalTPS(long totalTime) {
+        long totalTxns = (long) numRecords * numFiles;
+        logger.info(
+                "Total transactions: {}, time: {} sec, TPS: {}",
+                totalTxns,
+                totalTime / MILLISECONDS,
+                totalTxns * MILLISECONDS / Math.max(totalTime, 1));
+    }
+
     /**
      * Emulates crypto transfer.
      * Reads a batch of "account" pairs and updates them by transferring a random amount from one to another.
@@ -122,7 +137,8 @@ public abstract class CryptoBench extends VirtualMapBench {
         tps = BenchmarkMetrics.registerTPS();
         initializeFixedAccounts(virtualMap);
 
-        long prevTime = System.currentTimeMillis();
+        long startTime = System.currentTimeMillis();
+        long prevTime = startTime;
         for (int i = 1; i <= numFiles; ++i) {
             // Generate a new set of unique random keys
             final Integer[] keys = generateKeySet();
@@ -176,6 +192,7 @@ public abstract class CryptoBench extends VirtualMapBench {
             updateTPS(i, curTime - prevTime);
             prevTime = curTime;
         }
+        totalTPS(System.currentTimeMillis() - startTime);
 
         // Ensure the map is done with hashing/merging/flushing
         final VirtualMap finalMap = flushMap(virtualMap);
@@ -221,7 +238,8 @@ public abstract class CryptoBench extends VirtualMapBench {
 
         initializeFixedAccounts(virtualMap);
 
-        long prevTime = System.currentTimeMillis();
+        long startTime = System.currentTimeMillis();
+        long prevTime = startTime;
         for (int i = 1; i <= numFiles; ++i) {
             // Generate a new set of unique random keys
             final Integer[] keys = generateKeySet();
@@ -291,6 +309,7 @@ public abstract class CryptoBench extends VirtualMapBench {
             updateTPS(i, curTime - prevTime);
             prevTime = curTime;
         }
+        totalTPS(System.currentTimeMillis() - startTime);
 
         // Ensure the map is done with hashing/merging/flushing
         final VirtualMap finalMap = flushMap(virtualMap);
@@ -343,7 +362,8 @@ public abstract class CryptoBench extends VirtualMapBench {
 
         initializeFixedAccounts(virtualMap);
 
-        long prevTime = System.currentTimeMillis();
+        long startTime = System.currentTimeMillis();
+        long prevTime = startTime;
         for (int i = 1; i <= numFiles; ++i) {
             // Generate a new set of unique random keys
             final Integer[] keys = generateKeySet();
@@ -402,6 +422,7 @@ public abstract class CryptoBench extends VirtualMapBench {
                     map[FIXED_KEY_ID2] += 1;
                 }
             }
+            totalTPS(System.currentTimeMillis() - startTime);
 
             virtualMap = copyMap(virtualMap);
 
@@ -420,5 +441,14 @@ public abstract class CryptoBench extends VirtualMapBench {
             finalMap.release();
             finalMap.getDataSource().close();
         });
+    }
+
+    public static void main(String[] args) throws Exception {
+        final CryptoBench bench = new CryptoBench();
+        bench.setup();
+        bench.beforeTest();
+        bench.transferPrefetch();
+        bench.afterTest();
+        bench.destroy();
     }
 }
